refactor and revise collection metadata code
