+++ ui/src/components/collection/ViewCollection/Shell.component.vue
@@ -10,22 +10,7 @@
         <div class="px-6" v-if="data.userIsPermitted">
             <el-tabs v-model="data.activeTab" @tab-click="updateRouteOnTabSelect">
                 <el-tab-pane label="Collection Metadata" name="metadata">
-                    <describo-crate-builder
-                        v-if="data.activeTab === 'metadata' && data.metadataComponent.display"
-                        v-loading="data.metadataComponent.loading"
-                        :style="{ height: metadataPanelHeight }"
-                        class="overflow-scroll"
-                        :crate="data.crate"
-                        :profile="data.profile"
-                        :lookup="lookup"
-                        :enable-context-editor="false"
-                        :enable-crate-preview="true"
-                        :enable-browse-entities="true"
-                        :purge-unlinked-entities-before-save="true"
-                        @ready="data.metadataComponent.loading = false"
-                        @save:crate="saveCrate"
-                    >
-                    </describo-crate-builder>
+                    <MetadataComponent v-if="data.activeTab === 'metadata'" />
                 </el-tab-pane>
                 <el-tab-pane label="Collection Members" name="members">
                     <view-collection-members-component v-if="data.activeTab === 'members'" />
@@ -50,17 +35,13 @@ import AdministrationComponent from "./Administration/Shell.component.vue";
 import CollectionMembersComponent from "./CollectionMembers.component.vue";
 import ViewCollectionMembersComponent from "./ViewCollectionMembers.component.vue";
 import PublishComponent from "./PublishComponent/Shell.component.vue";
-import { ref, reactive, onMounted, onBeforeMount, inject, watch, computed } from "vue";
+import MetadataComponent from "../../Metadata.component.vue";
+import { reactive, onMounted, onBeforeMount, inject, watch, computed } from "vue";
 import { useRouter, useRoute } from "vue-router";
-import { useStore } from "vuex";
 import { ElMessage } from "element-plus";
-import { Lookup } from "../../lookup.js";
-import { debounce } from "lodash";
-const $store = useStore();
 const $route = useRoute();
 const $router = useRouter();
 const $http = inject("$http");
-const lookup = new Lookup({ $http });
 
 let props = defineProps({
     identifier: {
@@ -73,15 +54,7 @@ let data = reactive({
     routeWatcher: undefined,
     tabs: ["members", "metadata", "associate", "administration", "publish"],
     activeTab: "metadata",
-    metadataComponent: {
-        display: false,
-        loading: false,
-    },
     collection: {},
-    crate: {},
-    profile: {},
-    debouncedLoad: debounce(load, 300),
-    isAdmin: $store.state.user.administrator,
 });
 onBeforeMount(async () => {
     await checkUserAccess();
@@ -89,10 +62,6 @@ onBeforeMount(async () => {
 onMounted(() => {
     data.routeWatcher = watch(() => $route.path, updateRouteOnNav);
     updateRouteOnNav();
-    if (data.activeTab === "metadata") loadMetadata();
-});
-const metadataPanelHeight = computed(() => {
-    return `${window.innerHeight - 140}px`;
 });
 
 async function checkUserAccess() {
@@ -106,8 +75,6 @@ async function checkUserAccess() {
     data.userIsPermitted = true;
 }
 function updateRouteOnNav() {
-    data.metadataComponent.display = false;
-    if (data.activeTab === "metadata") loadMetadata();
     if (!$route.name.match(/^collections/)) {
         data.routeWatcher();
         return;
@@ -120,50 +87,6 @@ function updateRouteOnNav() {
     }
 }
 function updateRouteOnTabSelect(tab) {
-    data.metadataComponent.display = false;
     $router.push(tab.paneName);
-    if (tab.paneName === "metadata") loadMetadata();
-}
-async function loadMetadata() {
-    await new Promise((resolve) => setTimeout(resolve, 400));
-    data.metadataComponent.display = true;
-    await data.debouncedLoad();
-}
-async function load() {
-    if (!$route.meta.type && !$route.params.identifier) return;
-    data.metadataComponent.loading = true;
-
-    // load the crate file
-    let response = await $http.get({
-        route: `/describo/rocrate/${$route.meta.type}/${$route.params.identifier}`,
-    });
-    if (response.status !== 200) {
-        ElMessage.error(`Unable to retrieve RO Crate file`);
-    }
-    data.crate = (await response.json()).crate;
-
-    // load the profile
-    response = await $http.get({
-        route: `/describo/profile/${$route.meta.type}`,
-    });
-    if (response.status !== 200) {
-        ElMessage.error(`Unable to retrieve profile`);
-    }
-    data.profile = (await response.json()).profile;
-}
-async function saveCrate(data) {
-    let response = await $http.put({
-        route: `/describo/rocrate/${$route.meta.type}/${$route.params.identifier}`,
-        body: { data },
-    });
-    if (response.status !== 200) {
-        ElMessage.error(`Unable to save RO Crate file`);
-    }
-    // console.log("SAVE CRATE", JSON.stringify(data, null, 2));
-}
-function saveTemplate(template) {
-    // console.log("SAVE TEMPLATE", JSON.stringify(template, null, 2));
-    data.templates.push(template);
-    console.log(JSON.stringify(data.templates, null, 2));
 }
 </script>
